# performance improvement: optimize CRC computation